Move test to `poll_api`
